reorg pyterrier/java files to make more sense
