+++ pyterrier/java/__init__.py
@@ -1,8 +1,6 @@
 from pyterrier.java import mavenresolver
-from pyterrier.java._init import init, started, parallel_init, parallel_init_args, required, required_raise, before_init, autoclass, cast, legacy_init, JavaInitializer, CoreInit
-from pyterrier.java import config
-from pyterrier.java.config import configure
-from pyterrier.java.utils import redirect_stdouterr, bytebuffer_to_array, JavaClasses, J, add_jar, add_package, set_memory_limit, set_redirect_io, add_option, set_log_level
+from pyterrier.java.utils import JavaInitializer, init, started, parallel_init, parallel_init_args, required, required_raise, before_init, autoclass, cast, legacy_init, register_config, Configuration, JavaClasses
+from pyterrier.java.core import CoreInit, configure, redirect_stdouterr, bytebuffer_to_array, J, add_jar, add_package, set_memory_limit, set_redirect_io, add_option, set_log_level
 
 __all__ = [
     'add_jar', 'add_option', 'add_package', 'autoclass', 'before_init', 'bytebuffer_to_array', 'cast', 'config',
